+++ src/main/java/io/ejangs/docsa/domain/commit/app/create/CommitMySqlTxService.java
@@ -5,10 +5,10 @@ import io.ejangs.docsa.domain.commit.app.CommitQueryService;
 import io.ejangs.docsa.domain.commit.dto.request.CreateCommitRequest;
 import io.ejangs.docsa.domain.commit.entity.Commit;
 import io.ejangs.docsa.domain.commit.util.CommitMapper;
-import io.ejangs.docsa.domain.doc.app.EdgeService;
+import io.ejangs.docsa.domain.edge.app.EdgeService;
 import io.ejangs.docsa.domain.doc.entity.Doc;
-import io.ejangs.docsa.domain.doc.entity.Edge;
-import io.ejangs.docsa.domain.doc.util.EdgeMapper;
+import io.ejangs.docsa.domain.edge.entity.Edge;
+import io.ejangs.docsa.domain.edge.util.EdgeMapper;
 import io.ejangs.docsa.domain.save.app.SaveQueryService;
 import io.ejangs.docsa.global.mongo.deletion.dto.MongoIdsDto;
 import io.ejangs.docsa.global.util.RenewUpdatedAtHelper;
