+++ src/main/java/io/ejangs/docsa/domain/commit/dao/mysql/CommitRepository.java
@@ -1,6 +1,6 @@
 package io.ejangs.docsa.domain.commit.dao.mysql;
 
-import io.ejangs.docsa.domain.doc.dto.graph.CommitGraphDto;
+import io.ejangs.docsa.domain.edge.dto.graph.CommitGraphDto;
 import io.ejangs.docsa.domain.commit.entity.Commit;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,7 +11,7 @@ import java.util.List;
 public interface CommitRepository extends JpaRepository<Commit, Long> {
 
     @Query("""
-                SELECT new io.ejangs.docsa.domain.doc.dto.graph.CommitGraphDto(
+                SELECT new io.ejangs.docsa.domain.edge.dto.graph.CommitGraphDto(
                     c.id, c.branch.id, c.title, c.description, c.createdAt
                 )
                 FROM Commit c
