+++ src/main/java/io/ejangs/docsa/domain/edge/dto/graph/EdgeDto.java
@@ -1,4 +1,4 @@
-package io.ejangs.docsa.domain.doc.dto.graph;
+package io.ejangs.docsa.domain.edge.dto.graph;
 
 public record EdgeDto(
         Long from,
