+++ src/main/java/io/ejangs/docsa/global/config/JpaConfig.java
@@ -15,6 +15,7 @@ import org.springframework.transaction.PlatformTransactionManager;
         "io.ejangs.docsa.domain.branch.dao.mysql",
         "io.ejangs.docsa.domain.commit.dao.mysql",
         "io.ejangs.docsa.domain.doc.dao.mysql",
+        "io.ejangs.docsa.domain.edge.dao.mysql",
         "io.ejangs.docsa.domain.save.dao.mysql",
         "io.ejangs.docsa.domain.user.dao.mysql",
         "io.ejangs.docsa.global.mongo.deletion.dao.mysql"
