+++ src/test/java/io/ejangs/docsa/domain/commit/util/CommitMockTestUtils.java
@@ -6,7 +6,7 @@ import io.ejangs.docsa.domain.branch.entity.Branch;
 import io.ejangs.docsa.domain.commit.document.CommitBlockSequence;
 import io.ejangs.docsa.domain.commit.entity.Commit;
 import io.ejangs.docsa.domain.doc.entity.Doc;
-import io.ejangs.docsa.domain.doc.entity.Edge;
+import io.ejangs.docsa.domain.edge.entity.Edge;
 import io.ejangs.docsa.domain.user.entity.User;
 import java.util.HashMap;
 import java.util.List;
